Removed query runner as moving to new queryRunner app. Updated links to new app. Moved argument and testQuery components to queryDisplay directory.
