+++ src/router/methods/routes.ts
@@ -308,15 +308,6 @@ const routes: Array<RouteRecordRaw> = [
     ]
   },
   {
-    path: "/queryRunner",
-    name: "QueryRunner",
-    component: () => import("@/views/QueryRunner.vue"),
-    meta: {
-      requiresLicense: true,
-      requiresAuth: true
-    }
-  },
-  {
     path: "/codeGenerator",
     name: "CodeGenerator",
     component: () => import("@/views/CodeGen.vue"),
